[GECO-63] Addressing review comments given for controller
