+++ giles-eco/src/main/java/edu/asu/diging/gilesecosystem/web/core/service/reprocessing/IReprocessingService.java
@@ -1,0 +1,7 @@
+package edu.asu.diging.gilesecosystem.web.core.service.reprocessing;
+
+import edu.asu.diging.gilesecosystem.web.core.model.IDocument;
+
+public interface IReprocessingService {
+    public abstract void reprocessDocument(IDocument document);
+}
